datos correctos detail_incomes desde tree

+++ app/controllers/detail_incomes_controller.rb
@@ -1,7 +1,7 @@
 class DetailIncomesController < ApplicationController
  #GET /detail_incomes
     def index
-        @detail_incomes = DetailIncome.all.order('id DESC')
+        @detail_incomes= DetailIncome.where(detail_incomes: { tree_id: params[:tree_id] })
     end
 
     #GET /detail_incomes/new
@@ -15,7 +15,7 @@ class DetailIncomesController < ApplicationController
         @detail_income = DetailIncome.new(detail_income_params);
         if @detail_income.save
             flash[:notice] = "El ingreso de árbol ha sido creado con exito."
-            redirect_to :action => 'index'
+            redirect_to :action => 'index', :tree_id=>params[:detail_income][:tree_id]
         else
             render :new
         end
@@ -31,15 +31,20 @@ class DetailIncomesController < ApplicationController
         detail_income = DetailIncome.find(params[:detail_income][:id])
         detail_income.update_attributes(detail_income_params)
         flash[:notice] = "El ingreso de árbol ha sido editado con exito."
-        redirect_to :action => 'index'
+        redirect_to :action => 'index', :tree_id=>params[:detail_income][:tree_id]
     end
 
     # DELETE /detail_incomes
     def destroy
         detail_income = DetailIncome.find(params[:id])
-        detail_income.destroy
-        flash[:notice] = "El ingreso de árbol ha sido eliminado con exito."
-        redirect_to :action => 'index'
+        if ResourceActivitiesTree.where(tree_id: detail_income.tree).empty?
+          detail_income.destroy
+          flash[:notice] = "El ingreso de árbol ha sido eliminado con éxito."
+          redirect_to :action => 'index', :tree_id=>params[:tree_id]
+        else
+          flash[:notice] = "El ingreso de árboles tiene registros asociados."
+          redirect_to :action => 'index', :tree_id=>params[:tree_id]
+        end
     end
 
     private
